+++ example/src/App.tsx
@@ -1,11 +1,10 @@
-import React, { useRef, useState } from 'react'
+import React, { useState } from 'react'
 import {
   checkboxColumn,
+  Column,
   DataSheetGrid,
-  DataSheetGridRef,
   keyColumn,
   textColumn,
-  Column,
 } from 'react-datasheet-grid'
 import './style.css'
 
@@ -16,7 +15,6 @@ type Row = {
 }
 
 function App() {
-  const ref = useRef<DataSheetGridRef>(null)
   const [data, setData] = useState<Row[]>([
     { active: true, firstName: 'Elon', lastName: 'Musk' },
     { active: false, firstName: 'Jeff', lastName: 'Bezos' },
@@ -46,17 +44,7 @@ function App() {
         background: '#f3f3f3',
       }}
     >
-      <button
-        onClick={() => ref.current?.setActiveCell({ col: 'firstName', row: 1 })}
-      >
-        hello
-      </button>
-      <DataSheetGrid
-        ref={ref}
-        data={data}
-        onChange={setData}
-        columns={columns}
-      />
+      <DataSheetGrid data={data} onChange={setData} columns={columns} />
     </div>
   )
 }